+++ CHANGELOG.md
@@ -25,6 +25,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.15.3] - 2022-04-22
+
+### Fixed
+
+- Fixed bug in ``sat status`` which caused a traceback when certain component
+  state fields were missing in HSM.
+
 ## [3.15.2] - 2022-04-21
 
 ### Fixed